Simplify ArrayBindingPattern & Elision
Replace the three different ArrayBindingPatternN node types with a
single ArrayBindingPattern type, which has a list of bindings and
elisions, as well as a rest field.

Change Elision nodes so that they each only represent one comma, instead
of N commas.

In both cases we use a different grammar for these constructs than is
given in the spec, but the productions we use match exactly the same set
of inputs, so this does not break anything. The benefit is a
simplification of the AST.