Remove flags for now
We only had flags for a few productions (function, generator, and class
declarations) and this is a small subset of all the productions where we
are ultimately supposed to use them. To simplify the transition to an
explicit stack appraoch to parsing, remove these for now. We'll come
back later and add them in everywhere they need to be.