JSValue and Completion
The two most foundational data types used in the ECMAScript spec are
values and completions. Values represent all data created and used by
the program at runtime. Completions indicate the flow of control that
should be followed, such as returning a value or throwing an exception.

There are seven types of values: undefined, null, boolean, number,
string, symbol, and object. Each of these has a separate class, all of
which are subclasses of JSValue. For JSObject, the internal slots listed
in the spec are given as abstract methods, since there will ultimately
be many JSObject subclasses.

There are five types of completions: normal, break, continue, return,
and throw. These are also represented by separate classes, although
instead of sharing a common superclass, we define the Completion type as
the union of all five of these. NormalCompletion, as well as Completion
itself, are generic types in which the type parameter is that of the
value to be passed back. All the other types of completions can pass
back an arbitrary value.