Property keys and property descriptors
Add classes for data and accessor property descriptors, which are used
as entries in the properties dictionary maintained for each object.

Define an abstract class JSPropertyKey, from which both JSString and
JSSymbol inherit, since both strings and symbols can be used as property
keys.

For actually representing these keys in an object's property dictionary,
we use strings, so both these classes contain a stringRep() method which
converts the value into a form that can be used as such.

For symbols, we use "symbol:<id>", where <id> is a unique numeric
identifier generated each time a new JSSymbol object is created. For
strings, we use the form "string:<value>", where <value> is the actual
string. The prefix in the latter case is necessary to avoid the case
were a string key starting with "symbol:" is incorrectly interpreted.