Use dependency injection for JSObject methods
The JSObject class represents "ordinary objects" (as distinct from
"exotic objects", which have one or more non-default behaviours). In
order to implement these, we will need to make use of various functions
provided by the runtime modules, which themselves have a dependency on
datatypes.ts.

Unfortunately, due to the way in which typescript implements
inheritance, a module defining the superclass must be fully loaded
before the module defining the subclass. Since we'll be implementing
subclasses of JSObjects in some of the modules required to implement its
method, this would result in a circular dependency. Thus we need to keep
datatypes.ts independent of these. This is why datatypes.ts does not
import any other modules.

In order to provide implementations of methods in the JSObject class, we
now define these in a separate file, 09-01-ordinary.ts. These functions
are accessed via an exported object from RealmImpl, which is the
concrete class we use to implement the Realm interface defined in
datatypes.ts. To access these functions, all JSObject methods call
through to the functions accessible via the realm.