Use a custom map class for object properties
Previously, the proerties field of a JSObject was simply a native
JavaScript object on which we would set properties. However, using
regular objects as a map is problematic when some of the properties you
may not have actually added are actually present in the native object's
prototype; "toString" is an example.

With our own map class, we solve two problems:

1. We now explicitly record *only* those entries that have actually been
set, and do not inherit anything from a prototype.

2. The separation into a distinct class allows us, in the future, to
replace it with a lower-level implementation (e.g. a hash table) for
which we can have the compiler generate native code.

The class is called GenericMap and has a type parameter, as there are
multiple places in the spec where a map is needed. We do not use ES6's
Map class, as we ultimately need to use our own code for everything, can
we can't implement Map in terms of itself.