Fix ambiguity between property keys and strings
Object properties are identified by either strings or symbols,
collectively known as property keys. The set of all strings and the set
of all symbols are distinct. To enable both types to be used as key in a
map, we need to indicate for each entry which type it is. Until now,
we've used strings verbatim, and symbols of the form "symbol:<id>",
where <id> is unique numeric identifier for each symbol.  The problem
with this is that the symbol representations overlap with strings.

To properly disambiguate the two types of property keys, we now prefix
string names with "string:", and symbol names as above. Object propertes
are now stored in the PropertyKeyMap class, which is based on GenericMap
but uses JSPropertyKey instances as keys. However, adding the "string:"
prefix introduced problems where strings were in some cases being used
as both a property key representation and a declarative environment
name, the latter of which are strings, not property keys.

Our solution to this ambiguity is to split the Reference class into
multiple subclasses, where it can be determined from the type of
reference whether the name component is a string or property key. We now
have the following three classes, which inherit from AbstractReference:

- UnresolvableReference:

  Used when base is undefined. The name is a property key.

- PropertyReference:

  Used when base is an object, boolean, string, symbol or number. The
  name is a property key. PropertyReference also has a subclass,
  SuperReference.

- EnvironmentReference

  Used when base is an EnvironmentRecord. The name is a string.

The Reference type is now a union of the above three. This allows us to
use TypeScript's flow analysis to cover all three cases, testing for two
of the types and handling the third in an else clause. This is done in
GetValue, which uses the appropriate type of name for lookup, thus
avoiding treating property key representations prefix with "string:" in
lookups to environment records, as was initially happening when this
prefix was introduced.