Convert generic AST nodes to typed AST nodes
The parser produces "generic" nodes, which can have arbitrary names and
children. This is simpler from the parsing point of view, as it allows
us to define our grammer in a fairly abstract manner, but the results
are not readily suited to use in type checking or other parts of
compilation. For doing so, it is more useful to have one class per type
of AST node, so we can define and override methods to handle various
interpretation and compilation operations.

These classes were already defined (and in fact generated directly by
the parser in an earlier version of the codebase). This commit adds
functions which take the generic AST nodes (instances of GenericNode,
GenericStringNode, or GenericNumberNode) and creates a new tree with the
appropriate class based on the name. Each of the typed classes, in turn,
has properties of the appropriate types for its children, so we can
access these in a type-safe manner in methods that we write in the
future.