Work around compiler bug in control flow analysis
When the type of an object property is checked in an if statement, the
inferred type is narrowed appropriately in the true and false branches.
Unfortunately, this approach is too eager in the case where a property
may change as a result of mutations made by other functions called
between the point at which the check was made, and the point at which
the property is subsequently accessed.

TypeScript 2.1 extends this narrowing to primitive literal types, which
caused problems with the following code:

    if (p.cur == ".") {
        p.next();
        while ((p.cur != null) && (p.cur >= "0") && (p.cur <= "9"))
            p.next();
    }

generating the error:

    TS2365: Operator '>=' cannot be applied to types '"."' and '"0"'

In this example, p.next() causes p.cur to have a different value to what
it had before, but the compiler doesn't realise this. So when we access
p.cur after the call, it incorrectly assumes that the value is still
".". This demonstrates that this narrowing strategy is unsound, though
there are both pros & cons, as discussed here:

    https://github.com/Microsoft/TypeScript/issues/9998

The TypeScript team actually ran into this issue in their own code, in
the exact same context (examining the current token during lexical
analysis). @ahejlsberg's recommendation was to use functions in this
case instead of variables/properties (see his first comment on the above
issue). In this commit we do that; we call p.current() and assign the
result to a local variable, which the compiler detects modifications to.