Replace some untion types in AST with superclasses
Instead of using union types to refer to sets of classes that all belong
in the same place (e.g. items in a statement list), have the classes
that used to be members of those unions inherit from a common base class
where it makes sense to do so. This is done to simplify implementation
of methods for static and runtime semantics such as boundNames() where
those have to be present on all of the relevant subtypes, since we can
declare abstract methods on the superclass.

The inheritance relationships introduced in this commit are:

LexicalDeclarationNode (replaces ClassElementType)
    LetNode
    ConstNode

VarBindingNode (replaces VariableDeclarationListItemType)
    VarIdentifierNode
    VarPatternNode

StatementListItemNode (replaces StatementListItemType)
    StatementNode
    DeclarationNode

CaseClauseListItemNode (replaces CaseClauseListItemType)
    CaseClauseNode
    DefaultClauseNode

ImportListItemNode (replaces ImportListItemType)
    ImportSpecifierNode
    ImportAsSpecifierNode

ExportsListItemNode (replaces ExportsListItemType)
    ExportNormalSpecifierNode
    ExportAsSpecifierNode

ClassElementNode (replaces ClassElementType)
    MethodDefinitionNode
    StaticMethodDefinitionNode
    EmptyClassElementNode