Fix access to properties of the global object
When initializing the execution context for a module, set the outer
lexical environment to the global environment of the realm. Also, use
NewGlobalEnvironment instead of NewDeclarativeEnvironment for creating
the latter.

This fixes the ConsoleLog09.js test, which was failing due to an error
referencing "undefined", which is a property of the global object.