Fixes for TypeScript 2.2
TypeScript 2.2 introduces additional checks in relational comparisons,
reporting the following error if one of the operands is nullable:

    TS2531: Object is possibly 'null'

There was one place in the code where we were comparing the results of a
function call, and even though we had done the null check before (to
ensure the subsequent call would be non-null), this was not picked up
the compiler as it can't know whether another call to the function will
return now. So now we just assign the result to a local variable, which
it can safely check.