Fix modification of existing properties
ValidateAndApplyPropertyDescriptor takes a 'current' descriptor, which
is passed in when called from OrdinaryDefineOwnProperty. At the end of
the former, when it has completed all of its validation checks, it then
is supposed to update the relevant property descriptor on the object.

The bug here came from the fact that OrdinaryDefineOwnProperty passed a
*copy* of the property descriptor, rather than the original itself. So
modifying 'current' did nothing to the descriptor that was already
there. As a result, you could not modify the value of an existing
property once it had been set.

Now, instead of using current, we get the original property descriptor
and modify that. This ensures the new value (and other attributes)
persist.