Parser: Make execute() a wrapper method
In each of the Action subclasses, change the execute() method to
executeImpl(). All calls still go to execute(), but this is now
implemented only in Action, and acts as a wrapper for the impl method.
This is to us to add extra code that executes before and after every
action regardless of its type.