Add package.json, gulp build script + tslint
I've been avoiding this for a long time because ultimately I don't want
to rely on any third-party dependencies, as the VM should eventually
become self-hosting. However for the time being it's useful to rely on
tools like gulp and tslint for development.

Note that all dependencies are listed in package.json as devDependencies
as the intention being to keep the project always free of *runtime*
dependencies.