Partial fix for grammar dump() functions
This makes them closer to being consistent with the source, taking into
account the recent whitespace changes as well as the fact that actions
like pos and whitespace are now functions, not constants.

The only places where the output differs (once source comments are
ignored) is where a child node is displayed on the same line. This is
due to the stats() and space() functions being used as a prefix to each
node's dump() output.