Grammar: transform() method
This commit implements a mechanism for performing transformations on a
grammar based on variation of the visitor pattern. A call to a Grammar
object's transform() method with a transformer function will cause that
function to be invoked for the each top-level action node, which is
supplied as a parameter along with a transformation context object. The
transformer function should in turn call action.transform(ctx) to
recurse through the child action nodes.

If a given call to the transformer function returns the same Action node
it was passed, then no transformation has taken place and the resulting
subtree will be identical in the transformed grammar. This would always
be the case in a "visitor" use case, where the visitor (transformation
function) uses side effects to produce output or modify its own data
structures, and does not alter the tree.

If a given call to the transformer function returns a *different* Action
node to what it was passed, then that subtree will appear in place of
the original in the resulting grammar. Because Action nodes are
immutable, new versions of all ancestor nodes such as SequenceAction and
ChoiceAction will be created. This is only done when one or more
children different; otherwise, the nodes are shared.

The derive() method of Grammar can be used to add productions, whose
name can optionally be based on that of an exiting one. Transformations
are always performed on a *copy* of the Grammar object, which is
returned from the top-levle transform() call, so the newly-added
productions will be present only on the copy, not the original. The
context object passed to the transformation object includes a reference
to the new Grammar object, which should always be the one on which
define() is called.